vision/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    make compile-requirements
#
annotated-types==0.7.0
    # via pydantic
mypy==1.15.0
    # via -r vision/requirements.in
mypy-extensions==1.0.0
    # via mypy
pydantic==2.10.6
    # via -r vision/requirements.in
pydantic-core==2.27.2
    # via pydantic
pyyaml==6.0.2
    # via -r vision/requirements.in
ruff==0.9.7
    # via -r vision/requirements.in
types-pyyaml==6.0.12.20241230
    # via -r vision/requirements.in
typing-extensions==4.12.2
    # via
    #   mypy
    #   pydantic
    #   pydantic-core
